+++ sdk/azure-ai-assistant/azure/ai/assistant/management/assistant_client.py
@@ -9,10 +9,10 @@ from azure.ai.assistant.management.conversation_thread_client import Conversatio
 from azure.ai.assistant.management.conversation_thread_config import ConversationThreadConfig
 from azure.ai.assistant.management.exceptions import EngineError, InvalidJSONError
 from azure.ai.assistant.management.logger_module import logger
+
 from typing import Optional
 from datetime import datetime
 import json, time
-import copy
 
 
 class AssistantClient(BaseAssistantClient):
@@ -403,6 +403,7 @@ class AssistantClient(BaseAssistantClient):
                 thread_id=thread_id,
                 assistant_id=self._assistant_config.assistant_id,
                 instructions=self._assistant_config.instructions,
+                additional_instructions=additional_instructions,
                 event_handler=StreamEventHandler(self, thread_id, timeout=timeout),
                 timeout=timeout
             ) as stream:
@@ -534,25 +535,3 @@ class AssistantClient(BaseAssistantClient):
         except Exception as e:
             logger.error(f"Failed to update assistant with ID: {assistant_config.assistant_id}: {e}")
             raise EngineError(f"Failed to update assistant with ID: {assistant_config.assistant_id}: {e}")
-
-    def _update_tools(self, assistant_config: AssistantConfig):
-        tools = []
-        logger.info(f"Updating tools for assistant: {assistant_config.name}")
-        # Add the retrieval tool to the tools list if there are knowledge files
-        if assistant_config.knowledge_retrieval:
-            tools.append({"type": "retrieval"})
-        # Process and add the functions to the tools list if there are functions
-        if assistant_config.selected_functions:
-            modified_functions = []
-            for function in assistant_config.selected_functions:
-                # Create a copy of the function spec to avoid modifying the original
-                modified_function = copy.deepcopy(function)
-                # Remove the module field from the function spec
-                if "function" in modified_function and "module" in modified_function["function"]:
-                    del modified_function["function"]["module"]
-                modified_functions.append(modified_function)
-            tools.extend(modified_functions)
-        # Add the code interpreter to the tools list if there is a code interpreter
-        if assistant_config.code_interpreter:
-            tools.append({"type": "code_interpreter"})
-        return tools
